+++ src/RxQuery.js
@@ -428,3 +428,7 @@ export function create(op, queryObj, collection) {
 
     return ret;
 }
+
+export function isInstanceOf(obj) {
+    return obj instanceof RxQuery;
+}